+++ packages/core-js/modules/es.parse-float.js
@@ -1,8 +1,8 @@
 var $ = require('../internals/export');
-var parseFloatImplementation = require('../internals/number-parse-float');
+var $parseFloat = require('../internals/number-parse-float');
 
 // `parseFloat` method
 // https://tc39.es/ecma262/#sec-parsefloat-string
-$({ global: true, forced: parseFloat != parseFloatImplementation }, {
-  parseFloat: parseFloatImplementation
+$({ global: true, forced: parseFloat != $parseFloat }, {
+  parseFloat: $parseFloat
 });