+++ packages/core-js/modules/es.parse-int.js
@@ -1,8 +1,8 @@
 var $ = require('../internals/export');
-var parseIntImplementation = require('../internals/number-parse-int');
+var $parseInt = require('../internals/number-parse-int');
 
 // `parseInt` method
 // https://tc39.es/ecma262/#sec-parseint-string-radix
-$({ global: true, forced: parseInt != parseIntImplementation }, {
-  parseInt: parseIntImplementation
+$({ global: true, forced: parseInt != $parseInt }, {
+  parseInt: $parseInt
 });