+++ tests/pure/es.number.parse-float.js
@@ -19,5 +19,6 @@ QUnit.test('Number.parseFloat', assert => {
   /* eslint-disable es/no-symbol -- safe */
   if (typeof Symbol === 'function') {
     assert.throws(() => parseFloat(Symbol()), 'throws on symbol argument');
+    assert.throws(() => parseFloat(Object(Symbol())), 'throws on boxed symbol argument');
   }
 });