+++ tests/pure/es.parse-float.js
@@ -18,5 +18,6 @@ QUnit.test('parseFloat', assert => {
   /* eslint-disable es/no-symbol -- safe */
   if (typeof Symbol === 'function') {
     assert.throws(() => parseFloat(Symbol()), 'throws on symbol argument');
+    assert.throws(() => parseFloat(Object(Symbol())), 'throws on boxed symbol argument');
   }
 });