+++ tests/pure/es.parse-int.js
@@ -35,5 +35,6 @@ QUnit.test('parseInt', assert => {
   /* eslint-disable es/no-symbol -- safe */
   if (typeof Symbol === 'function') {
     assert.throws(() => parseInt(Symbol()), 'throws on symbol argument');
+    assert.throws(() => parseInt(Object(Symbol())), 'throws on boxed symbol argument');
   }
 });