+++ tests/tests/es.number.parse-float.js
@@ -21,5 +21,6 @@ QUnit.test('Number.parseFloat', assert => {
 
   if (typeof Symbol === 'function' && !Symbol.sham) {
     assert.throws(() => parseFloat(Symbol()), 'throws on symbol argument');
+    assert.throws(() => parseFloat(Object(Symbol())), 'throws on boxed symbol argument');
   }
 });