+++ tests/tests/es.number.parse-int.js
@@ -38,5 +38,6 @@ QUnit.test('Number.parseInt', assert => {
 
   if (typeof Symbol === 'function' && !Symbol.sham) {
     assert.throws(() => parseInt(Symbol()), 'throws on symbol argument');
+    assert.throws(() => parseInt(Object(Symbol())), 'throws on boxed symbol argument');
   }
 });